+++ app/src/main/java/com/tassadar/lorrismobile/joystick/JoystickListener.java
@@ -2,5 +2,5 @@ package com.tassadar.lorrismobile.joystick;
 
 
 public interface JoystickListener {
-    public void onValueChanged(int ax1, int ax2);
+    public void onValueChanged(double ax1, double ax2);
 }
